pyproject.toml
[tool.poetry]
name = "soli-api"
version = "0.1.0"
description = "API for SOLI - the Standard for Open Legal Information"
authors = ["ALEA Institute <hello@aleainstitute.ai>"]
homepage = "https://soli.openlegalstandard.org"
repository = "https://github.com/alea-institute/soli-api"
license = "MIT"
readme = "README.md"
include = ["soli_api"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
soli-python = {version = "^0.1.4", extras=["search"]}
fastapi = "^0.112.2"
uvicorn = "^0.30.6"

[tool.poetry.group.dev.dependencies]
sphinx = "^7.4.7"
black = "^24.4.2"
myst-parser = "^3.0.1"
sphinx-book-theme = "^1.1.3"
sphinxcontrib-mermaid = "^0.9.2"
pytest = "^8.3.1"
pytest-asyncio = "^0.23.8"
pytest-cov = "^5.0.0"
pylint = "^3.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
max-line-length = 120
disable = [
    "line-too-long",
    "too-few-public-methods",
    "no-self-argument",
    "cyclic-import"
]
ignore-patterns = ["test_.*?.py", "conftest.py"]

[tool.pylint.messages_control]
max-args = 10
max-attributes = 15
max-branches = 15
max-locals = 25
min-public-methods = 0

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
addopts = "--cov=soli_api --cov-report=term-missing --cov-report=xml"
